Backend/refactor/merging-dasboard-coin-event-types
