+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Types/Coins/CoinHistory.hs
@@ -11,16 +11,15 @@
 
  the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
 -}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Domain.Types.Coins.CoinHistory where
 
-import qualified "dashboard-helper-api" API.Types.ProviderPlatform.Management.DriverCoins as DCoins hiding (CoinStatus)
-import Data.OpenApi (ToSchema)
+import qualified "dashboard-helper-api" Dashboard.ProviderPlatform.Management.DriverCoin as DCoins
 import Data.Time
 import EulerHS.Prelude hiding (id, state)
 import Kernel.Types.Id
 import qualified Lib.DriverCoins.Types as DCT
-import Tools.Beam.UtilsTH (mkBeamInstancesForEnum)
 
 data CoinHistory = CoinHistory
   { id :: Id CoinHistory,
@@ -33,12 +32,8 @@ data CoinHistory = CoinHistory
     updatedAt :: UTCTime,
     expirationAt :: Maybe UTCTime,
     coinsUsed :: Int,
-    status :: CoinStatus,
+    status :: DCT.CoinStatus,
     bulkUploadTitle :: Maybe DCoins.Translations,
     entityId :: Maybe Text
   }
   deriving (Generic, FromJSON, ToJSON, Show)
-
-data CoinStatus = Used | Remaining deriving (Show, Eq, Ord, Read, Generic, ToJSON, FromJSON, ToSchema)
-
-$(mkBeamInstancesForEnum ''CoinStatus)
